Debug get_np_results
Now that we are using iter and not p_dots throughout, the object
being drawn from has type dataframe and not list.
Therefore matching is exact for column names now,
breaking previous partial matches (to
var_exp_s in this case, instead of var_exp_spline).
The fact that var_exp_s was used is likely unintentional.